+++ src/packages/core/modal/modal-route-registration.ts
@@ -1,4 +1,4 @@
-import { UmbModalContext } from './modal-handler.js';
+import { UmbModalContext } from './modal-context.js';
 import { UmbModalConfig, UmbModalManagerContext } from './modal-manager.context.js';
 import { UmbModalToken } from './token/modal-token.js';
 import type { IRouterSlot } from '@umbraco-cms/backoffice/external/router-slot';
